# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/teachme-api/public/api",
    "paths": {
...
        "/aulas": {
            "get": {
                "tags": [
                    "Aula"
                ],
                "summary": "Get a listing of the Aulas.",
                "description": "Get all Aulas",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Aula"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Aula"
                ],
                "summary": "Store a newly created Aula in storage",
                "description": "Store Aula",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Aula that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Aula"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Aula"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/aulas/{id}": {
            "get": {
                "tags": [
                    "Aula"
                ],
                "summary": "Display the specified Aula",
                "description": "Get Aula",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Aula",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Aula"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Aula"
                ],
                "summary": "Update the specified Aula in storage",
                "description": "Update Aula",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Aula",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Aula that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Aula"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Aula"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Aula"
                ],
                "summary": "Remove the specified Aula from storage",
                "description": "Delete Aula",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Aula",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Aula": {
            "required": [
                "CdUsuario_Aluno",
                "CdAnuncio",
                "Horario"
            ],
            "properties": {
                "CdAula": {
                    "description": "CdAula",
                    "type": "integer",
                    "format": "int32"
                },
                "CdUsuario_Aluno": {
                    "description": "CdUsuario_Aluno",
                    "type": "integer",
                    "format": "int32"
                },
                "CdAnuncio": {
                    "description": "CdAnuncio",
                    "type": "integer",
                    "format": "int32"
                },
                "Horario": {
                    "description": "Horario",
                    "type": "string"
                }
            }
        },
...
    }
}